Wrong check permissions for tab manage attachments in profile

+++ components/com_kunena/controllers/user.php
@@ -446,21 +446,23 @@ class KunenaControllerUser extends KunenaController {
 		return true;
 	}
 
-	function delfile() {
+	public function delfile() {
 		if (! JRequest::checkToken ()) {
 			$this->app->enqueueMessage ( JText::_ ( 'COM_KUNENA_ERROR_TOKEN' ), 'error' );
 			$this->redirectBack ();
 		}
 		$cids = JRequest::getVar ( 'cid', array (), 'post', 'array' );
 
-		$number = count($cids);
+		if ( !empty($cids) ) {
+			$number = count($cids);
 
-		foreach( $cids as $id ) {
-			$attachment = KunenaForumMessageAttachmentHelper::get($id);
-			$attachment->delete();
+			foreach( $cids as $id ) {
+				$attachment = KunenaForumMessageAttachmentHelper::get($id);
+				$attachment->delete();
+			}
+
+			$this->app->enqueueMessage ( JText::sprintf( 'COM_KUNENA_ATTACHMENTS_DELETE_SUCCESSFULLY', $number) );
+			$this->redirectBack ();
 		}
-
-		$this->app->enqueueMessage ( JText::sprintf( 'COM_KUNENA_ATTACHMENTS_DELETE_SUCCESSFULLY', $number) );
-		$this->redirectBack ();
 	}
 }
